Move event types to events.d.ts
